backend/requirements.txt
# AI Social Media Automation - SaaS Web App
# ==========================================

# Backend Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
python-dotenv==1.0.0
openai==1.12.0
google-generativeai==0.4.0
httpx==0.26.0
apscheduler==3.10.4
jinja2==3.1.3
python-multipart==0.0.6
python-dateutil==2.8.2

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Encryption for API keys
cryptography==42.0.2

# Database
psycopg2-binary==2.9.9  # PostgreSQL
aiosqlite==0.19.0       # SQLite async support

# Email validation
email-validator==2.1.0

# CORS
starlette==0.35.1


